CTS changes for MemoryFile constructor throwing IOException.
This change updates the CTS tests for MemoryFile
to catch IOExceptions thrown by the MemoryFile constructor.

Before, the MemoryFile constructor could throw IOException, but did not
declare this fact, see http://b/issue?id=1881829
This has now been fixed in
https://android-git.corp.google.com/g/2618
Unfortunately that change must be submitted together with this one in order
not to break the build.