Add shared uid app install test to cts.
This host test uses hosttestlib to verify that an apk which declares
a shared user id fails to install if its signed with a different certificate
than an existing app that declares the same shared userid.

Included in this CL are:
 - InstallTests.java, which is the test itself
 - two empty test apps that declare the same user id
 - two test certificates

The change to actually include this test in the CTS distribution will be
forthcoming in a separate CL, once b/1781345 is implemented.

For now, the test can be run via the 'runtest cts-appinstall' command.