CTS: Use plain /usr/bin/python for the build instead of /usr/bin/python2.4
Even though the python style guide explicitly requests that
python2.4 should be used in the shebang line, there is no such
executable or link on the Mac build machine.

Changing the shebang to /usr/bin/python should do the trick. The
popular '/usr/bin/env python' is not allowed according to the style
guide due to issues with killall.