Bug 2194752
Fix some test errors in android.database cts tests..

Some of the tests cases explicitly orphan some SQLiteStatements
that have been compiled.  When the tests are cleaned up, closing the
database would throw a SQLiteException that there are some unfinalized
statements.

Now the the tests that leave these orphaned statements explicitly
close the database, and eat that particular exception.  The database
file will be unlinked, and a new one is created for the next test,
so if the file is not close cleanly, this shoudln't be a problem

Ideally, SQLiteDatabase.removeSQLiteClosable() would be accessible by
unit tests, to clean up better, but this would either require the
package that the unit tests belong in to change or to make that
method public