CTS tests for CharSequence list+array in Bundle+Intent
This tests the methods added in Id82686f6ca8c9501f6db8a07018278a78ddacd05

I also added a Parcel roundtrip to most of the tests, since that's
what Bundles are mostly used for. Without that, the tests are prety trivial,
since they just test that the HashMap in the Bundle is not insane.

See http://b/issue?id=2468093

Change-Id: Ia945b96f29a3014df4685531c94da8956443f3ab