DO NOT MERGE - Speculative fix for intermittent CTS UI test failures.
UI related tests fail if the keyguard is currently displayed.
Previously, the CTS harness used a separate util app which would disable the
keyguard upon receiving a broadcast.

However, if the util app's process is killed by the system, the keyguard is
immediately displayed. Its not known for sure, but this could explain the
intermittent UI test failures.

This commit adds logic to disable the keyguard in the CTS test runner itself.

Bug 2342725.

Change-Id: I87659e5e3547296dc0e4bd549c979aef325fc90e