Fix ConfigurationTest#writeToParcelTest
Bug 2455171

Update ConfigurationTest's writeToParcelTest that has fallen out
of date with changes to Configuration. It still passes on the
emulator, because it was testing the defaults which are all zeros.
Change the test case a bit to test alternating zeros and non-zero
constants to make it fail when things get out of date. Furthermore,
test null and non-null Locales too.

Change-Id: Idfbb8a200cf3e5f73209d3381a40c6bbeebb6102