Fix CtsTestServer Socket Deadlocks
Bug 2558977

When one thread was blocking on socket.accept() and another thread
was trying to call socket.close(), there would be a deadlock that
would cause the WebView tests to hang. Instead of having another
thread call socket.close(), issue a request to a shutdown url that
the socket.accept() thread will recognize and subsequently call
socket.close() in the same thread to fix the issue.

Change-Id: I9152b5bf39893998ea44fcb31e18699652f94e57