Increase cts-app Test Stability
Bug 2639189

Many cts-app tests start LaunchpadActivity which in turn calls
startActivityForResult and expects onActivityResult to setResult
to SUCCESS. Other tests don't use startActivityForResult and are
coded to depend on onDestroy to send back a result which defaults
to cancelled! When ActivityManagerService decides to change the
configuration for some reason, onDestroy terminates and never
gets back the result. It ends up setting cancelled and causes
the tests to fail! Thus, fix LaunchpadActivity to ignore
configuration changes that ActivityManagerService to prevent
it from being destroyed.

Change-Id: Iea2944f66b4dd1ecdfccfcc4bcdc9354de2eb623