Further Activity Lifecycle Testing Improvements
Bug 2639189 2663328 2189784

The prior change ran the lifecycle checking activity and if it
failed to match ran the activity again with a different expected
lifecycle. However, running the activity again could cause the
activity to follow the original lifecycle that failed. Thus,
follow the expected lifecycle as long as possible and then switch
to other possible lifecycles when running into lifecycle mismatches.

Also do not check the lifecycle for onStop or onDestroy methods,
because those are not guaranteed to be executed at all. Delete tests
that checked onStop and onDestroy being called, because they will
probably hang in the future.

Change-Id: I6f721d70861cd9b22523b8aa9b883e7336daab22