CTS Test for ARM ABI Version
Bug 2713199

Use the native cpufeatures API from the NDK project to make sure
that if the CPU is v7 capable then it should be reported by the
Build.CPU_ABI constant. Tested by changing the TARGET_CPU_ABI
variable in passion-common's BoardConfigCommon.mk to armeabi
and then making sure the test caught it.

Change-Id: Iedc5ab663dd90a2e7be2508b3a3a7641b04e1bb4