WebViewTest Delayed Check for savePicture
Bug 2814482

The file saving is done in a separate thread in WebView, so wait a
bit when checking for the file's size.

It probably should be noted in the JavaDoc that this is an
asynchronous call. Furthermore, the method never returns false even
if the file saving failed, because the saving is done in a Runnable
that will fail silently...

Change-Id: Ied27fbb0b09f8fd7090cbf7392d81d81a9b91ec5