Fix WebViewTest#testAccessCertificate Assertion
Bug 2841873

setCertificate was being called twice once with a null and then with
a non-null certificate. The first call though set the
mHasCalledSetCertificate and then this caused the later not-null
assertion to fail. Get rid of those methods and just directly
use setCertificate and getCertificate to check that the certificate
is set properly.

Change-Id: I6710daa77e42a95f22b8ed9e7d8633fdb93e7f33