Test for Checking OpenGL ES Version
Bug 2845550

Detect what version of OpenGL should be reported and check that
the system property matches. If the device doesn't support 2,
make sure the property is 1 or returning an undefined value
which implies 1. If the device supports 2, set a GLSurfaceView
to 2.0 and check that the version string makes sense. By
default the version should be 1.0 before setting the GLSurfaceView
to 2.

Change-Id: I240f30fb2a19b5e168bb5ba0c0da23268895b31d