Some Assorted WebKit Test Fixes
Bug 2888212

The tests call a "loadUrl" method that waits until getProgress
returns 100. However, it appears that the updates returned by
getProgress are not accurrate unless there is a WebChromeClient
set on the WebView as seen in the code of
CallbackProxy#onProgressChanged. Setting the client should
fix the flakiness of the tests.

Call setInitialScale in WebSettingsTest so that the testFindNext
test will have to scroll to find the next instance of the word
"all" since the default scaling in QVGA will zoom out to show
two cases of "all" which won't require scrolling.

Change-Id: I3e313794e11f303b26ebd6dc1cde2720774ab439