CtsVerifier Test Result Infrastructure
Add a TestResult class that allows test activities to set their
finishing results as either pass or fail. Reflect the pass or
fail status in the TestListActivity's test list. This is done
via a ContentProvider that reads and writes to a database with
the TestListAdapter listening for changes.

Use the new API in the SUID Files test as an example. The other
activities should be able to be easily adjusted to use the API.
Just call TestResult.setPassedResult or .setFailedResult during
some time in the activity before you finish.

Change-Id: Ia6a1f75e2ac9d03cb22ed233bedf392e3fb28f0e