Eclair CTS Test for ARM ABI Version
Bug 2923901

This is the Eclair version of the ARMEABI test added to Froyo
for http://b/2713199. Here is what needed to be tweaked to make
it work:

1. I didn't see an open NDK Eclair project, so I copied the
   cpufeatures source code to a temporary directory under CTS.

2. There was no public SDK Java method to get the second ABI
   property. However, I went ahead and used the hidden API
   due to the importance of the test.

This test already exists in Froyo and uses the proper public API.
This test will not be merged into Froyo and future releases.

Change-Id: I20fcc21096f468fb733fcf34842b218e41dabbc1