Add "Omitted" State for CTS Tests
Add a new "omitted" state for tests. This is different from "not
executed" to fix a problem where CTS would keep retrying to run
the not executed tests over and over again.

In a prior change, I switched the default test result from pass to
not executed in the test result parser, since I didn't want skipped
tests to be reported as passed (or tests that failed horribly to
report pass). This worked fine when running a specific package "-p"
but started to fail when running the entire plan "--plan CTS." This
is because there are two code paths where one infinitely retries
if there are any tests not marked as "not executed" but the other
just retries once. Tests with the new annotations thus caused an
infinite loop of retrying... By setting tests to "omitted" by
default, the tests won't be retried over and over.

Change-Id: I3b4995089605860b1ed5e6b752dde811132f16f8