Fix opcode tests for new_array and new_instance
Some of these tests were trying to access a private inner class
of TestStubs to trigger an IllegalAccessError. However, the tests
kept throwing an NoClassDefFoundError instead. This seems to be
because the classes.dex does not have the inner classes due to
the setup done by BuildDalvikSuite. To get around this, I changed
the byte code to try to access a package protected class instead
to trigger the IllegalAccessError...

Bug 3147582

Change-Id: Icc4ce84011af19617d5fd6244171707a1b1c7b3c