Make CookieSyncManagerTest use stricter cookie format.
This test is failing with the Chrome HTTP stack because Chrome is
fairly strict about cookie handling, but the test expects the system
to allow some non-standard behaviour.

First, Chrome was rejecting the cookies because they weren't being
set on a complete URL. Fixed by calling getFileUrl() to ensure they
have a valid "file:" prefix.

Second, Chrome was accepting the "a = b" cookie but returning it as
"a=b", which is the format required by the RFC. I've updated the test
to use "a=b" throughout.

Bug: 3241780

Change-Id: I69a9d200c94df00a45f4db07d7350dbb8ab7ee6a