Fix GridViewTest#testSetHorizontalSpacing
The test creates a GridView with 20 columns and puts 3 elements into it.
It then sets the horizontal spacing to values like 0 and 5 and checks
that the spacing between the views are 0 and 5. However, the GridView is
distributing the elements on the wide screen, so the spacing isn't
correct. Thus, set the stretch mode to none to fix the issue. This isn't
a new behavior in GridView.

Change-Id: Id30d378021b8de9d6c1e87cf79d5e2b75467749a