Delete Useless LayoutTests
Bug 3198659

Couple new tests were failing because getLineDirections in the test
class was returning null. It turns out there is no way via the public
API to construct a Directions object. Its been like this for a while,
so there were a ton of other tests checking to see that NPE occurred.
Destroy all the tests that do so...

Change-Id: I31f6c162e42085fb1742c1573099d8f9908e2dc1