Fix test to pass in a reasonable timestamp.
As of Honeycomb, the system checks the timestamp of events before
delivering them so that it can drop stale events from the input queue.
Some DialogTests were injecting motion events with a timestamp of 0
which is usually going to be more than 10 seconds in the past so they
were dropped.

Other CTS tests do the right thing and pass in reasonable timestamps.

Bug: 3350093
Change-Id: I3335947e257d977488882b2ab1b4babfbb80e014