Fix 2 MediaStore Tests
These tests were passing in files that did not really exist and sending
a fake size to the ContentProvider. The provider no longer blindly
trusts the size given by the ContentValues. Since it actually opens the
file and checks its length, the size it really inserted didn't match the
fake value passed via the test. Thus, actually create a fake file of a
certain length and use that in the test. The update tests following the
insertion still blindly trust the size value in the ContentValues, so
those don't need an actual file to back the test...

The providers also insert their own date modified times rather than
trusting the given ContentValues, so make sure to use seconds as the
documentation recommends rather than milliseconds.

Change-Id: I3a1fcee296b5831e08a75f1a6fddd276b9055c07