Fix WindowTest Tests
Removed all the methods trying to test the options menu. There is code
that deliberately no-ops these methods if the app is targeted for HC.
Ideally, we should have another CTS stub with a lower target SDK version
and then test the options panel there.

Finally, fix an assertion that no longer held due to the System Bar
taking up part of the screen.

Change-Id: I2637e1494623e4c2dd7cda0f8d2d3d030dc1d356