Fix ContextTest and ContextWrapperTest Interaction
Bug 3188260

ContextWrapperTest#testAccessTheme passed when run on its own
but failed when executed in batch mode. ContextTest was
interfering by calling obtainedStyleAttributes which was
calling getTheme. getTheme sets the theme to the default
if there is no theme. You can only set the theme once in
the context, so future calls like those in testAccessTheme
no longer worked and broke the test.

These ContextWrapperTests are somewhat silly since the
implementation of ContextWrapper is to just call the
same methods in the context it wraps...so I moved the
test into ContextTest and made sure to set the theme
to the test theme at the beginning.

These ContextTests are also funny, because there are multiple
implementations of Context and this one just happens to work
with the Context given back by AndroidTestCase...

Change-Id: I710439ea09b8c048178faacb269157d4ba4341cb