Fixing race in android.hardware.cts.CameraTest#testZoom
Remove the checking process that compares "given zoom value
from callback" and "gotten zoom value via getZoom()".

Because there is a race condition in that the callback happens
and before the call to getZoom the zoom has already been updated.

When "onZoomChange()" is executed during smooth zoom,
the return value of "getZoom()" in checking process
won't correspond to the argument "value"
in case that the zoom has been already updated before the
call to it.

Change-Id: Ie9c277c09efb1510d1314b26652815d04a79be4f