Fix for CTS testScreenConfiguration
The previous implementation of testScreenConfiguration() failed to
pass some acceptable aspect ratios.
Example: A screen in 16:9 screen ratio that is 480 pixels wide,
would have to be 853,333 pixels high to be exactly 16:9. This is
of course not possible, so we must allow both 853 and 854 pixels
in this example. The previous implementation would only accept
853.

Change-Id: Ic5e7fa523869d1036821bb29c602ee82ec750238