Update VelocityTracker tests.
Modified the test so that all of the samples are at least 10ms
apart.  This is because the new VelocityTracker does not attempt
to compute the velocity when the time difference is too small to
give accurate results.

The computed values are also slightly different because the new
VelocityTracker does not discard the most recent sample.  Instead,
it is more careful about how it handles pointers going down or
up to ensure that the pointer's last known velocity is accurately
reported.

Change-Id: Ie551d343ed0be22cb5b87d5a16bf4bb74210c44a