CTS AudioTrackTest: Changing audiotrack buffer size from minBuffSize to 2*minBuffSize
This patch modifies the audio track buffer size for some test cases.
If the buffer size is not multiple of framesize in track.write, the audio track write function
hangs as frame count * framesize doesn't match the buffer size sent.
eg: if minbuffersize = 7524, buffersizesent = minbuffersize/2 = 3762

framecount = 3762/4(frame_size) = 940.5 (as it is fractional number considers as 940)

framecount*frame_size = 940*4=3760 and the write function hangs to get the remaining 2 bytes.

As buffersize is always multiple of frame_size, it is good to sent the entire buffer to track.write
function to avoid this kind of issues.

Change-Id: I6fda86ad12dc3b1e1735caa3bb527de87c7b3586
Signed-off-by: karimuddin sayed <sayed.karimuddin@ti.com>