CTS: Fixes for android.hardware.cts.CameraTest#testCancelAutofocus
- setPreviewDisplay() was omitted after release().
release() disconnects and releases the Camera object resources.
If a camera uses an overlay for preview, the overlay object is
also destroyed when release() is used. At the same time,
the camera documentation states that a call to either
setPreviewDisplay() or setPreviewTexture() is necessary in
order to start the preview.

Change-Id: Ic77e8a5ea952414c610850241be9e09018b02b20
Signed-off-by: Julian Shandorov <jshandorov@mm-sol.com>