Revert "CTS: Fixes a possible "out of memory" exception"
This reverts commit 7eade134e4c8fdaee5b7dfb0446dc90085b0408a.

This causes the test to fail on Nexus S, so I'm reverting
the change for Gingerbread. However, it will be included in
a future CTS release.

Additional info from Wu-cheng:
"The change made the test faster. If apps call setParameters
right after jpeg callback is received, it will fail. This is
not a regression and I believe it does not impact users."

Change-Id: I14ce1235581dac5543983f60ed047ae9331a754a