Manual Gyroscope Test for CTS Verifier
This manual test consists of 6 stages. Each stage asks the tester
to rotate the device along an axis in a certain direction. The
activity turns green when the device is reporting the expected
values and red when it is reporting unexpected values.

Finally, there is a warning when the sensor values are higher
than expected and may be in degrees.

Change-Id: I1b653f00d2375cf34f225023c663d824dba0eec2