Wait for completing the initialization of WebViewCore
This findAll() will fail when WebView$mNativeClass field is not updated.
That field is updated after the initialization of WebViewCore.
The test may execute before the completion of it. Because the test case
waits for the completion of it on UI thread, the test cannot notice
the completion message of it. In that case, test will fail.
So the test should wait for the completion of init on non-UI thread.

Change-Id: I7ad811d2d8259e58afd78cae9736dbf78df547e6