Fix for bug 4442572: CookieSyncManagerTest fails when executed with other tests ...but passes by itself
The test tries to set a cookie on a file url. From build HMI42
file cookies are disabled by default.

It works when the test is running alone as cookies for file url's
are enabled in the start of the test. It doesn't work when all
tests run together as the setting only works when it is set before
the cookie manager is created. When several tests run after each
other the CookieManager is reused.

Change-Id: If1ab3a75cc4f2f969a6f21523f3e53dd45ec5e7a