Add some strings for screen where the text stayed in
When the text - "Scrolls the specified widget to the specified
coordinates, except constrains the X scrolling position to the
horizontal regions of the text that will be visible after scrolling
to the specified Y position." - has stayed in the screen, it is not
scrolled. Therefore, some strings is added to make the text scroll.

Change-Id: I149f3d89ba60266527a08cf5fd3562f22272a304