It is neccessary to keep a reference to ParcelFileDescriptor
When obtainDescriptor() returns back to test methods,
ParcelFileDescriptor object is not reachable from them.
So it might be destroyed by gc if a memory was low.
ParcelFileDescriptor would close a real file descriptor
when it was finalized. In that case, bitmap related tests will fail.
(Because FileDescriptor$descriptor is gone! i.e. "-1")

To avoid this fail, the test methods keep the reference to
ParcelFileDescriptor object in them.

Change-Id: I3980179b0545086e522fd327e265211c1f6d1d53