Add more comprehensive unit tests for WebP Image.
Added test for end-to-end WebP encoding and decoding, by encoding a PNG
image to WebP and decode WebP image back to Bitmap and compare against
the PNG BitMap.
Note for integrator: The file 'res/drawable/webp_test.webp' is already
present in the internal Android-git repo, but not present in AOSP (as
AOSP is somehow not in sync with internal Android-git repo w.r.t
'platform/cts.git' project). I had to add this file to make this test
complete w.r.t it's dependencies.

Change-Id: I68cfa71290797aee1d5645be0deec0a3bc1fe727