CTS Verifier USB Accessory Test
Bug 5001945

This test is an adaptation of the AccessoryChat and
accessorychat USB accessory testing programs found in
frameworks/base/libs/usb/tests. It requires starting
up the CTS Verifier, connecting a device, and then
running the cts-usb-accessory program.

The accessorychat is now called cts-usb-accessory.
I removed the condition that limited accessorychat
to connect to only a certain vendor's device. Furthermore,
you do not type messages to send a message. The program
merely responds with an acknowledgement message that the
Android device receives and logs.

AccessoryChat was melded into the CTS Verifier, the manual
tester. It also does not allow a user to enter custom
messages, but it sends a message to start a dialog with
the cts-usb-accessory. After 10 or messages are collected,
the test activates the pass button.

Finally, there is an additional variable in CtsTestCaseList
to add cts-usb-accessory to the CTS tools directory. There
will need to be another change in build/core/tasks/cts.mk
to make sure all the executables specified by the variable
are copied over properly.

Change-Id: I79add3fa169d28de206b7ed2de97cc8904592765