CTS Verifier Data Backup Test
Bug 2749924

Add a test activity that shows some values for preferences and files.
It has a button to populate these preferences and files with random
values. Request the user to run several adb commands to execute the
backup process, unistall the app, and then verify that the random
values are still present. This is a pretty loose test but better than
nothing I suppose...

Also add support for backing up the current test results...

Change-Id: I8f0555c97cc74efc4c082735733c01c7d136e790