Fix AutoText setting of locale
Setting the locale in AutoTextTest does not update
the configuration properly. This problem appears
only when the default locale is not set to English
(property persist.sys.language).

The test just changes the default locale used for
initialization, but other locale fields remain not updated.
This fix also updates the configuration so that all
locale dependent fields are consistent.

Change-Id: I7d4c1d12a87e546a3320926dfd212d299263d844
Signed-off-by: Irina Tirdea <irina.tirdea@intel.com>