Add EMMA-coverage runner for Calendar provider
To get EMMA test coverage information for the Calendar provider CTS
test, we need a way to tell the provider to dump its coverage state
when the CTS test finishes.  To that end, we define an alternate
instrumentation test runner that issues special provider queries
at the start and end of the testrun.

Bug 2361048

Change-Id: I9130506366a194626eddd98eaccac5e1fc85b471