Add HTTP/RTSP streaming tests to CtsVerifier.
RTSP tests are temporarily disabled while network issues are worked out.

Also, this refactors the class hierarchy (in particular, for test Activities) to remove
some duplicated code, and introduces the concept of a Test ID appended to the class name
instead of merely using the class name, allowing us to use the same class for multiple
tests.

Change-Id: I0a9feedaa92c23f1bf366ca16ccae672ad7e783f