Fix Suid File Scanner Activity
I am pretty sure this used to work, but it seems the SUID scanner
test result isn't recorded since the only option you get is to
use the back button after the test completes. I added some
pass fail buttons to the activity. Created a general layout that
can be reused in the future.

Change-Id: I6f12fd4d46130045a6ca8f69a8a7d6644fa2a916