Include Audio Quality Details in XML Report
Bug 5352891

This will add a <details>Log from test...</details> for tests that
pass an optional string when setting the test result or from
overriding the getTestResult method.

This is implemented by adding another column to the test results
database. I modified the backup agent to backup this column as well.
I removed a line in the backup agent that marked the test to fail
due to a mismatched backup key. I kept getting the key
"../databases/results.db" and I think that is from some older
backup that I have... I confirmed backup still works fine though.

Change-Id: Iffec1853d8e1dcbcf9feda821716d071860be069