Remove a test case for mms intent-filter
This test case will make sure a existence of activities for the
implicit intent which has the type of "vnd.android.cursor.dir/mms".
In AOSP, the intent-filter such as above is declared by Mms
application. However it isn't included in core applications
defined by CDD. So CTS shouldn't have that test case for all devices.

Change-Id: Ib8a322cd7b62a561767cced69bea69958a877128