Don't Call pass() Twice in SearchManagerTest
Bug 5444914

In one of the tests, pass() was being called by the handler's doStep
method and then again after checknig the doStep's return value.
The first pass caused the activity to terminate and clean up its
references. Then the handler's pass call would get a NPE when it
finally executed.

I also cleaned up a lot of junk in the corresponding test.

Change-Id: I65611bdda03e3243a646cd0a51b11e6986f4b019