New Holo Theme Tests
Bug 5404694

These tests check that the Holo theme assets have not been
altered. The code is based upon the techniques in android.theme
of inflating a view in a fixed dp with layout, taking a bitmap,
and comparing it with a reference bitmap.

It does not incorporate the activity snapshot or split activity
snapshot tests of android.theme due to test instability issues.

This version tries to simplify a couple things:

- A utility activity is included in CtsThemeTestCases that let
  you view individual tests and lets you generate all the reference
  images.

- Saves the bitmaps to the external storage directory, so you don't
  need root!

- Doesn't generate landscape permutations of the bitmaps, because
  the widgets shouldn't differ too much by being in landscape...

This initial change includes assets for hdpi, large-mdpi, and xhdpi.

Change-Id: I9d9f304494ee95bf8d772f1c0e0e0b73a8fc8317