Fix ConfigurationScreenLayoutTest
Bug 5739739

The test fails when the screen is locked. The disable keyguard
code isn't invoked, because this test package doesn't use
InstrumentationCtsTestRunner. Rather than depending on
InstrumentationCtsTestRunner and having to move this
activity to com.android.cts.stubs, I added code to disable the
keyguard which is simpler than the one InstrumentationCtsTestRunner.

Change-Id: Iaacba8407d6fc8615d2e87c4c45fca9e76515107