Fix AbsListViewTest#testGetFocusedRect
The test was recently updated to have a longer adapter to support
larger screen devices. However, the adapter is longer than one
screen, so now selected items are scrolled to the top which causes
testGetFocusedRect to fail. Thus, create another adapter that
has only a couple of items that shouldn't fill the screen to restore
the test.

Change-Id: I6b17675ac74ddde036567754f10eb0ab67b44d77