added cts tests for audio preprocessing
Added cts tests for noise suppressor, acoustic echo canceler and
automatic gain control.

The APIs for these audio pre processing effects are simple and we just need to
test that creating and enabling/disabling the effect works.
These effects are not mandatory and the test shall pass if they are not avaible
on the tested platform.

Change-Id: Ice8f1dd43733d14dd0bd204628c8ddee6f21e044