Fixed AnimationDrawable isRunning() logic
A fix in AnimationDrawable for bug 5782773 meant that
the existing AnimationDrawableTest now fails, because it was checking
for state which is incorrect. Basically, if you setVisible() on an
AnimationDrawable, it *should* be running; the previous bug was that
it was not marked as running until the next animation frame.

This CL reverses the logic for the isRunning() test.

Issue #6294015 AnimationDrawableTest#testSetVisible fails on JRN04B

Change-Id: I606d1cf20be65d34f6b8cbf2be04b42378ed940b