FileSystemPermissionTest: Don't stop early
When scanning the filesystem looking for bad files /
directories, the current tests stop at the first bad one.
This suppresses other findings, and gives an inaccurate
state of the security of the device.

Collect all bad files and directories before we fail, and
print them all out as part of the failure message.

Change-Id: I8494694765034b5fd557cd2618b8ddad86a17a68