Add manual camera preview format tests to CtsVerifier.
Display output from preview callbacks in all formats supported by the
device, allowing the tester to determine if the format is properly
produced, and that the mandatory formats are present.

Currently supports the mandatory YV12 and NV21 formats, and the YUY2
format. Selecting other formats will display a dialog stating the
format is not testable through this test.

Bug: 6330501
Change-Id: I1ad1447fc96b9bd6c4ef498c14d55b2e02de07f6