Get rid of zergrush reflection test
The ZergRush vulnerability is no longer present in
this Android release. No need to test for it further.

This test is failing due to commit dc410cb3499683509770bb88cb771a91acd16bd2
in frameworks/base. The MOUNT_UNMOUNT_FILESYSTEMS permission is
signature|system, and is no longer available to third party apps.

Bug: 6448705
Change-Id: Ibf7a8439e5faef4e4c73bc3977c56a71dd93ff30