Fix android.cts.text.TextUtilsTest#testEllipsize test
This fixes the CTS test regression tracked in bug 7292719, by fixing a
flaw in the test.

Make the textWidth computation more robust in the case where the test
string could contain a kerning pair with the ellipsis. The new Roboto
font update does indeed kern (r, ellipsis), which is valid.

Change-Id: I12eedcc3c24496a96efd8a56bd92234ddbdcfc52