Make the cts test record a bit longer
The initial delay of the recording increases from 700 ms to 1000 ms, due to the change
of some audio assets. As a result, recording 1000 ms may end up recording nothing; hence
the expected stop failure.

This patch increases the recording duration from 1 second to 3 seconds, and changed some
hardcoded values to use constants for easier maintence in the future.

o related-to-bug: 6971305

Change-Id: I9642087fad74c696ee5882c25892095ed89ac62f