Remove useless code.
The function
  protected void sendSms(PendingIntent sentIntent, PendingIntent deliveryIntent,
           String currentNumber)

is never called, making this class essentially identical to
the class it extends (NoReceiveSmsPermissionTest).

It seems like the original author intended this method to be:

  protected void sendSms(String currentNumber, String text, PendingIntent sentIntent,
          PendingIntent deliveryIntent) {

(note the different arguments) which would then @Override the version from
NoReceiveSmsPermissionTest. However, if you refactor the code to make that
change, you essentially end up with identical definitions of sendSms.

Since refactoring would result in identical code, the best thing to do is
to delete this class.

This code was originally added in commit 7000873a066ca6f6e5c174b8f5fd944dd8c222b1

Change-Id: I90fb75acf88df7918f831e3e919516beac7c9fe6