Fix up WebViewTest#testLoadDataWithBaseUrl. DO NOT MERGE
Trivial cherry pick from master, see
I8fd1f65fb2db5163d128a580176ecb93b188cf95.

The Chromium WebView requires that the "history URL" is a valid
URL. Update the test so that it passes one.

Additionally, in Chromium the data url formed by the load specifies the
charset parameter, so we can't check for a comma ',' directly
after the MIME type like we did with the classic WebView.

Also, add a bit of extra logging to help track some flakiness in
this test.

Bug: 7907196
Change-Id: Ie2503da404ca2113c4f8efb0b02a74bc15a62ad8