Fix for bug 8631592 android.widget.cts.TextViewTest#testGetFocusedRect
The test was assuming that the coordinates coming back from
getPrimaryHorizontal() would always be an integer (so would match exactly
the int-valued coordinates of getFocusedRect), but this is not a valid
assumption when text is being laid out with higher precision. This change
loosens the test slightly so that only the integer component needs to
match.

Change-Id: I8d5e336242c6578a7e46068afbf2e5f337f9db40