Adding capability request for the helper accessibility services.
We added APIs for accessibility services to declare in advance their
capabilities so we can warn the user when enabling then. The capability
to enable touch exploration was the only old one and it required user
consent, i.e. it was dynamically granted. The test fails because the
target API version of the helper service it installs are targeting
API higher 18 which requires the new attributes. This change adds
the attribute to the services config files.

bug:8724789

Change-Id: Ib30b6cec14241a0812f0b5e07989a164312641e1