short-term fix for flaky CTS tests that need focus
A number of the tests in cts-text are flaky because a number of their
operations require window focus, but the frameworks give no guarantee
that an Activity has window focus at any given point in the Activity
lifecycle.

This commit addresses the issue by having the affected tests block in
setUp() until window focus has been obtained.

Bug: 9054507
Change-Id: I922c3649d7f378e7d3a7560937919f1b48713a59