Fix up WebViewTest.testFlingScroll for WebViewChromium. DO NOT MERGE
This makes the testFlingScroll test wait a bit after loading the page
(so that content size change notifications have time to propagate all
the way up to the UI thread) which is to prevent flakyness.
This also increases the size of the flingable area to work around an
unrelated overview-scale bug.

BUG:10287454
Change-Id: Ia53e5356ab86ef6668ad185fba75998d9296ed3a