Adding exceptions for File system permission test
Adding 2 exceptions which are needed by somedevices.

/dev/fimg2d

FIMG2D device is 2D acceleration H/W that is used by SKIA, HWC
and HDMI. And SKIA APIs are called by "user apps" via JNI. It
means fimg2d using buffer that is mapped by "user" and it's reason
why the device should have '666' permission. Additionally,
fimg2d use only user buffer and free from security problem.

/dev/mobicore-user

To be able to share memory between Android Client App and Secure
App the MobiCore drive must create a mapping table of the Android
Application and send it to the SWd OS.Since the user space memory
allocated by malloc is most likely fragmented in the physical
memory the MobiCore Linux driver needs access to the Client App's
task object and also to the buffer's start and length to be able
to build the translation table. The translation table is
essentialy a table of physical addresses the Linux Driver sends
to the SWd OS so it can make this memory available to the Secure
App. Because the only way to setup this shared buffer is by
getting the physical address of every 4KB page in the buffer we
need it done in Kernel Space and not in the daemon. We also have
a special case where a contigous buffer is allocated from kernel
space instead of using malloc in the client app.

Bug: 10428016

Change-Id: I72da71e2cd670a98bb06998e89db9e9f532fc7c5
Signed-off-by: Raj Mamadgi <rmamadgi@sta.samsung.com>