media.DecodeEditEncodeTest: Increase bitrate for QCIF test
Some devices fail the color test due to allocating too few bits on
the first I-frame.  While this can questionably be a device error
given that we encode the QCIF sample at 1Mbps, we do not specify
a quality requirement for encoders, and the color tolerance is also
somewhat arbitrary.

Change-Id: I9fc7e2acd8961cdb96f8cf90d9c4f7549ae114f7
Signed-off-by: Lajos Molnar <lajos@google.com>
Bug: 10503038