Fix DeviceAdminActivationTest broken by ProGuard.
ProGuard is now enabled on the CtsDeviceAdmin package used by
DeviceAdminActivationTest. As a result, the
CtsDeviceAdminActivationTestActivity.setOnActivityResultListener is
stripped by ProGuard and cannot be found by the test at runtime,
leading to a NoSuchMethodError.

This CL fixes the issue by marking this method with
@VisibleForTesting (an annotation from the Guava library). ProGuard
by default leaves methods marked with this annotation untouched.

Bug: 10664979
Change-Id: Idef2b92a219ddf4c9c8667bfd17b6bf60d499030