Update CTS tests: remove acquireContentProviderClient test.
This test is dependent upon whether or not the content provider defines
any path permissions - if it does the acquireContentProvider call will
not throw. Since acquiring the content provider alone is no the important
part of the test, remove it. Querying/updating the content provider is
handled by the other test cases here, and these are what matter.

Change-Id: I880d01fe25b69c705bfb81d0c76a225e5100a981
Signed-off-by: Mathew Inwood <mathewi@google.com>