FileSystemPermissionTest: Whitelist adsprpc-smd device
The underlying device driver allows for clients to
make remote method invocations between applications
and DSP processor. DSP processor's software stack also
allows 3rd parties to run their algorithm on OpenDSP platform.

Such android library/apps would load and use libadsprpc.so,
which opens /dev/adsprpc-smd driver. Without access to the
device node applications will not be talk to use OpenDSP apps.

This general concept is very similar to how graphics processing
is offloaded to the GPU. In this case DSP favorable workload is
offloaded to DSP and depending on the processing that is offloaded,
different apps/services would need to be able to interact with
DSP this way.

The device is considered secure because only signed and
authenticated libraries, signed by the device manufacturer can
run on DSP.  The objects that will be manipulating these
buffers on the remote processor will be signed and trusted
entities, thereby alleviating any fear of intentional
scribbling of these buffers.

The driver itself is also secure.

Bug: 11710243
Change-Id: I05259abcfba01ca04673bf30e8a7861567ea8320