java7: Fix usage of Process / ProcessBuilder.
The javadoc tool for Java7 is stricter than Java6, and when
invoked using ProcessBuilder / Process, the piped error stream
becomes full and the child process is blocked.

Work around this by redirecting the error stream to the
standard stream. Error output (javadoc errors due to missing
dependencies) will be filtered out anyway.

Also, replace a pointless use of Scanner with a BufferedReader.

Change-Id: Iac2cf804667f2223f182801d84446ab9f37124b0