Use runtime lookup of config_useFixedVolume in AudioManager
In the current code the config setting config_useFixedVolume was
transformed into the resource id at compile time. This cause errors
in the tests if this resource id was not used for the same config
setting in the device under test. This is solved by looking up
the resource id during runtime instead.

Change-Id: I5fed34af745b92ec88b527c853eb4996f8ed6115