Simplify RSUtils.java.
There's no need for the forEach methods.  The methods filling
arrays with random values should not care what the "step & skip"
is, as done previously, as entries that are skipped by our API
should just be ignored.  It's actually better to test with random
data in there, to make sure we don't somehow use the zero that
used to be there.

Change-Id: I7724b45f4a187699533a87a36c5a3d992092943a