Native tests use --gtest_list_tests for test list.
Rewrite the nativescanner to parse the output of --gtest_list_tests.

Modify the rules for the xml test list generation to build the special
host test list executables and pipe the output to the native scanner.

Modify all of the native tests to create a host list executable.

Remove the known failures list for bionic tests.

Bug: 10914944

Change-Id: Ib4d7da0f260eeef08db6fd083cfeecf92459979b