Native tests use a text file based test list.
Rewrite the nativescanner to parse the output of --gtest_list_tests.

Modify the rules for the xml test list generation to cat the text file
as the source of the test lists. The text file was generated by running
all three gtest tests with the --gtest_list_tests option and saving it
to a file.

Remove the known failures list for bionic tests.

Bug: 10914944

(partial cherry-pick of 2830d721b4173ac7a91e4fc8b9cc0cafaaf86460 from aosp master)

Change-Id: I0a35f191f050e007d2b657b0ad871f38a8f27a9b