Fix CTS android.provider package testAccountRemoval_hasDeleteLogsForContacts seldom fail
The test case ContentResolver.query() method will return cursor to application,
but application should close the cursor if it has no need to be reused longer,
 or cursor leak will happen, and what's worse, "Out of memory" issue will hit 
this application, finally application may be killed because of memory limit 
if long time running. So calling "close" method of cursor is necessary for 
the case.

Change-Id: I2745f2c3e3e054307fd1981530e7f9f0d75f4ced
Signed-off-by: Junjie Hu <junjie.hu@mediatek.com>