Fix a concurrency bug in OpenSSLHeartbleedTest.
The test was not guaranteed to observe all of the data transmitted by
the client in the scenarios where the client failed. As a result, the
test was occasionally failing (instead of passing) when the client
terminated the TLS/SSL handshake with the fatal alert
unexpected_message.

This CL fixes the test by ensuring that the test observes all of the
data exchanged between the client and the server.

Bug: 13906893
Change-Id: Ieb835821db4bb6403492ad99073294104722d539