CTS:Start a GL activity with a unreachable stopping criteria
The test case will start a GL activity and CTS thread monitors the latency of
the GL view through "dumpsys SurfaceFlinger --latency SurfaceView". If CTS
thread tries to dumpsys after GL activity finished 400 frames and exited, a
parser JAVA expectation will happen. In low-end device the designed is OK,
because the GL activity is hard to finish before the CTS thread to monitor it.
On the other hand, in some specify device, for example, the high-refresh rate
LCD phone,  the GL thread has higher probability to fail. This is un-reasonable.
We add the number of frame up to 1000 to avoid this situation.

Change-Id: If18539dd9489b542405a7d26f4a8bb038592fa94
Signed-off-by: Junjie.Hu <junjie.hu@mediatek.com>