Add okhttp tests as a CTS package
Being added in goog/internal to take advantage of the
work that has been done to support JUnit4.

The tested classes and tests are kept in their original package
(com.squareup.okhttp) because many of the tests rely on package
visibility to work.

For CTS tests the tests and code are loaded in different
ClassLoaders so that won't work. Instead, we test okhttp as a
standalone library.

Bug: 14462336
Change-Id: Ic00a5ccb134d406b3d1939172338d8833fdbb71b