Fix CTS Verifier for BLE: Reliable Write Fail
When run the test case for executing write,  the client side will execute the
following steps:
1. begin    2. write   3. execute write
After executing step 3, the test case still can't pass.

The root cause is the server side doesn't send the response when executing the
onExecuteWrite method.It'll cause the client side's test case fail.

To fix the problem, we must send the response in the onExecuteWrite method.

Change-Id: I4995d5bea0383b6a51769bddee3483832c89acf6
Signed-off-by: Junjie.Hu <junjie.hu@mediatek.com>