Skip encoding tests for unsupported resolutions.
Some low-tier chipsets do not support video encoding at resolutions
above 720p (1280x720). However, some of the CTS tests attempt to encode
at these resolutions even if they are not supported, and therefore fail.

To fix this, because running the test itself, check that the video size
is supported, and skip the test if not.

To check whether the resolution is supported, for the AVC case, compare
the requested resolution with the limits specified in the AVC
specification, based on the level supported by the encoder.

Change-Id: I3ecb8d30235c8fef854551ce2b10c65ae72d86f9
Signed-off-by: Yoongeun Kwon <yoongeun.kwon@lge.com>