NoReadLogsPermissionTest: don't depend on directory read access
In L, the /dev/log directory isn't world-readable. Expand out the
full list of files, and test each file individually to make sure
it has the appropriate permissions.

Allow for the non-existance of any file. With the migration to the
userspace logger, we expect these entries to go away over time.

Bug: 17594208
Change-Id: Id275e1aeb8e14ebd77ec53bc5a54bfb6cc7853fc