CTS: Add some known WebGL failures that can only be fixed in Chrome.
Whether empty/unset vertex position should be an error appears
to be ambiguous in GLES, so I don't think we can file vendor bugs.

Similarly I've verified that the failure in webgl_specific should
be handled by Chrome and not the driver, so there must be a Chrome
bug causing the different behavior on different devices.

Change-Id: I7b8c3d56d8f8ab0770cb9fa968aed2edd3d6e611