Add CTS test for Roboto fonts
As of L, the CDD now requires Roboto fonts for fidelity to the Material
Design theme. The Roboto 2 font also has significantly expanded Unicode
coverage including many characters from Unicode 7. This test verifies
the presence and appearance of the fonts, and also coverage, using
golden reference images. The threshold is tuned to allow for legitimate
variations in rendering.

Bug: 17109280
Change-Id: I7abbed6cad2ecbb98caa8da2f0f8c8d51a419a91