Make tests more reliable by cleaning up worker threads on cancelation. b/17888343
This change prevents false-negatives caused by a left-of sensor test running in the background
while a new test begins execution.
Most of the changes are one-liners to allow the propagation of InterruptedExceptions that were
swallowed deep in the sensor test platform.
The real changes to handle interruption are part of the base activity classes and the parallel
sensor operation.

Change-Id: Ida0abf8da55b9a03856a1d99a6b306915835ca3f