CameraITS: check the reported crop region
The reported crop region when both YUV and RAW are requested should
be the crop region for the YUV frame within tolerance.

Add the gating to skip test_crop_regions on devices that don't support
freeform cropping.

Fix the requesting crop region in test_crop_region_raw to be a supported
center region.

    Bug: 17994909

Change-Id: Iaeec3a8b51185e4f4a8084548252ad0c34e0aa5c