CTS Verifier: hide AID prefix reader test if no support
The reader side of the AID prefix test works also on devices
without AID prefix registration support. However to some
it seems confusing as this makes them look for the emulator
side tests.

Hence removing also the test on the reader side when the
device denies the capability via supportsAidPrefixRegistration()

Bug: 18147492
Change-Id: If0be6be1257956d768f64d65e5198a6f29882531