Tweak alarm batching CTS test
Introduce an intermediate exact alarm to specifically force earlier delivery
of the inexact windowed alarm on "fixed" versions of the OS where inexact
alarms are not permitted to batch together with exact ones.  The net change
here is to allow us to use a significantly stronger test for failure and
makes the test more robust to other alarm activity going on concurrently.

Bug 18726051

Change-Id: Ifff3bdd3dbda9393b07090f7a4cd3f602bc30513