Further refinement of the alarm batching test
The test now places an inexact alarm with a known window, and an
exact alarm 1ms before the end of that window.  Under the older,
problematic batching implementation these two were permitted to
coalesce; this is no longer the case.  The test repeats this scenario
several times and watches the delivery interval between the two alarms.
If that interval is very short more on more than 20% of trials, it
is assumed that the system is allowing the alarms to coalesce
and the test fails.

Bug 18726051

Change-Id: I5aa1e0f4ee8a0246488889171373381c48646cab