SELinuxHostTest:  Add a test for the AOSP seapp_contexts entries.
Add a test that requires the device seapp_contexts to be identical
to the general AOSP seapp_contexts entries.  Other entries may be
added after the AOSP entries.  Retain the separate checkseapp validation
test to ensure that all of the entries are valid and that none of the
later entries conflict with the AOSP entries.

Change-Id: I431c1ab22fc53749f623937154b9ec43469d9645
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>