Respect hard restriction to minimum password length in CTS
Internally a hard restriction has been introduced to minimum password length. Former
CTS tests failed because of this. Current CL ensures all test cases that are supposed
to succeed are at least as long as the minimum required length. Also extra assertions
have been added to check for failure in case a shorter password is set.

Bug: 20418344
Change-Id: Ie88939adb9ff0ad16493ba53e7ea046389f88f7e