Rewrite TextView onSaveInstanceState CTS tests
The previous tests directly access the SaveState parcel, and make
assumptions about its internal layout, which is not valid. This patch
replaces those tests with new ones that actually save and restore
instance state to a real TextView, using stable public interfaces.

The save state also contains state related to undo, but coverage for
that was already added.

Bug: 21209641
Change-Id: I9e9a3b37afb8356021e7b7232ee36823a0ce63d5