Simplify CTS testing for policies that apply to DO and PO
Currently, code for policies that apply to device and profile owner
situations has to be duplicated or added to a shared library. This
change introduces one app that can act as device and profile owner and
holds all common test classes. Added the runtime permission tests to
this app as a first example.

Bug: 22192363
Change-Id: I6a7a8bf9f64a44f5aed4d8f70c3868a603b98757