Add CTS conference test
Added a simple conference test and made some changes in the telecom
CTS framework to support this.
Changes in the CTS telecom framework:
1. Store the static instance of the CtsConnectionService that is
registered with telecom via the manifest. The Mock connection service
implementation in the various tests can use this CtsConnectionService
static instance to talk back to the telecom.
2. Add support for adding conference calls and splitting from conference
calls.

BUG: 20303674
Change-Id: I52decaec788fa81b8c4f0ebc34b600ab395f3e92