Refactor CtsVerifier BYOD tests to use existing classes
Eliminating custom Adapter and TestResult classes/enums. Also making it
possible to start activities that return com.android.cts.verifier.TestResult
and use those results to set the status of the test on the list of BYOD tests.
All the while retaining the capability to launch dialogs with 'pass' and 'fail'
buttons instead of launching instances of PassFailActivity.

Bug: 21512534
Change-Id: I32714e944249df942cf56d5bf6552d9b2bb31c20