Add RemoteConference API Cts tests.
For RemoteConference API testing, we create 2 remote connections on the
CtsRemoteConnectionService from CtsConnectionService and then create a
remote conference using these remote connections.
The conference on the CtsConnectionService is plumbed to the conference
created on CtsRemoteConnectionService via the RemoteConference object.

Other changes:
1. Refactored the Remote tests to have a separate base class.
2 .Some changes in RemoteConnectionTest & ConferenceTest due to some
framework changes needed to accomodate the RemoteConferenceTest.
3. Rebased to make this change on top of the CTS telecom tests framework
changes by Yorke.

BUG:20303674
Change-Id: Icd081dbb851d794101c1a4a1a19f239da0f76349