Add FileSystemPermissionTest exception for DRM-based GPU drivers.
On newer kernels, the 'card0' device is no longer the only device node
required to use a DRM-based GPU driver stack. The render node architecture
introduced in Linux 3.15 requires that 'renderD128' be world-writable.

(This issue affects the android-3.18 common kernel; android-3.14 and
 android-3.10 are not affected.)

This device node is used by any process using EGL/GLES, which includes system
processes such as SurfaceFlinger and most sandboxed native/java applications.
It is therefore not possible to use only group-writable permissions.

Imagination Technologies believe that the renderD128 device node is secure
(in the context of a PowerVR GPU stack only) and has been properly audited.

Change-Id: I35c440a3d83c6fc6cb045215aa2addb0866e15ed