Few stability fixes in Telecom CTS tests.
Fixing some failures when running all the telecom tests back to back
in a suite:
1. Ensure that the connection/incall services created by test are
unbound before exiting a test to ensure that each test cleans up the
telecom state.
2. In ConnectionService tests, we need to wait for the new calls
corresponding to the new connections created. Since the calls created
via |addExistingConnection| API in ConnectionService take a little while
to bubble upto inCallService, we might end up missing to cleanup this
new call as a part of test teardown and this might interfere with
subsequent tests.
3. Destroy the MockConnection object on reject/abort. This was resulting
in the ExtendedInCallService#testRejectIncomingCall leaving behind a
stale connection object.
4. Since the remote phone accounts are registered/enabled via ADB commands
(since they need system permission), we need to add a small wait to ensure
that the account has been successfully registered and available in
PhoneAccountRegistrar.
5. Use |creatTestNumber| method instead of |getTestNumber|  to create a
new unique phone number for each incoming call in the test.
6. Some other minor cleanup of unused code.

BUG: 23604254
Change-Id: Iff123f122f9cb43ec89f6c18cdb76e4b33dd4a94